+++ src/components/tabletInfo.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import {Table, Button} from 'react-bootstrap';
 
-export const ViewTablet =({users, editRow, handleOpen, setEdit})=>{
+export const ViewTablet =({users, editRow})=>{
     return(
         <Table responsive bordered striped >
             <thead>
                 <tr>
+                    <th>#</th>
                     <th>Nombre</th>
                     <th>Apellido</th>
                     <th>CI/RIF</th>
@@ -15,8 +16,9 @@ export const ViewTablet =({users, editRow, handleOpen, setEdit})=>{
             </thead>
             <tbody>
                 {users.length > 0 ? (
-                    users.map(user =>
-                        <tr key={user.id} >
+                    users.map((user, index) =>
+                        <tr key={index} >
+                            <td>{user.id}</td>
                             <td>{user.name}</td>
                             <td>{user.lastName}</td>
                             <td>{user.idDocument}</td>
@@ -25,18 +27,14 @@ export const ViewTablet =({users, editRow, handleOpen, setEdit})=>{
                                 <Button
                                     variant='outline-success'
                                     className='mx-1'
-                                    handleOpen={handleOpen}
-                                    onClick={() => { 
-                                        editRow(user)
-                                        setEdit(true)
+                                    onClick={() => {
+                                        editRow(user, true)
                                         }}>Editar</Button>
                                 <Button
                                     variant='outline-danger'
                                     className='mx-1'
-                                    handleOpen={handleOpen}
                                     onClick={() => {
-                                        editRow(user)
-                                        setEdit(false)
+                                        editRow(user, false)
                                         }}>Borrar</Button>
                             </td>
                         </tr>
